clip_ig/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
open_clip_torch==2.32.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.3.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchtyping==0.1.5
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typeguard==2.13.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
webdataset==0.2.111
xxhash==3.5.0
yarl==1.20.0
